+++ src/main/kotlin/icu/windea/bbcode/editor/BBCodeAnnotator.kt
@@ -5,8 +5,8 @@ import com.intellij.psi.*
 import com.intellij.psi.util.*
 import icu.windea.bbcode.*
 import icu.windea.bbcode.intentions.*
+import icu.windea.bbcode.lang.*
 import icu.windea.bbcode.psi.*
-import icu.windea.bbcode.util.*
 
 //com.intellij.codeInspection.htmlInspections.XmlWrongClosingTagNameInspection