+++ src/main/kotlin/icu/windea/bbcode/lang/schema/BBCodeTagType.kt
@@ -1,0 +1,14 @@
+package icu.windea.bbcode.lang.schema
+
+enum class BBCodeTagType {
+    Block,
+    Inline,
+    Line,
+    ;
+
+    companion object {
+        fun resolve(name: String): BBCodeTagType? {
+            return entries.find { it.name.equals(name, true) }
+        }
+    }
+}